Use Environment Variables for Configuration
The configuration for the quote server is currently hard-coded in the
main.go file. This is not ideal, as it makes it difficult to change the
configuration without recompiling the code.

This commit changes this behavior so that the connection strings are
read from an environment variable. This allows the configuration to be
changed on container startup without rebuild the container image.

Modified timo-quote-driver-src/main.go:
- Add ENVVAR: REDIS_CONNECTION_ADDRESS
  DEFAULT: quote-price-redis:6379

- Add ENVVAR: RABBITMQ_CONNECTION_STRING
  DEFAULT: amqp://guest:guest@rabbitmq:5672/

- Add ENVVAR: QUOTE_SERVER_ADDRESS
  DEFAULT: quoteserve.seng.uvic.ca:4444

Modified docker-compose.yml:
- Add environment variables to quote server. Values differ between
  development and production.